# FMS/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Force Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Member {
  id: number;
  /** @nullable */
  discordId?: string | null;
  discordUsername: string;
  /** @nullable */
  callsign?: string | null;
  /** @nullable */
  characterName?: string | null;
  /** @nullable */
  rankId?: number | null;
  /** @nullable */
  rankName?: string | null;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  joinedAt: string;
  /** @nullable */
  totalPatrolHours?: number | null;
}

export interface MemberInput {
  discordId?: string;
  discordUsername: string;
  callsign?: string;
  characterName?: string;
  rankId?: number;
  departmentId?: number;
  status: string;
  notes?: string;
}

export interface MemberUpdate {
  discordId?: string;
  discordUsername?: string;
  callsign?: string;
  characterName?: string;
  /** @nullable */
  rankId?: number | null;
  /** @nullable */
  departmentId?: number | null;
  status?: string;
  notes?: string;
}

export interface Rank {
  id: number;
  name: string;
  level: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  discordRoleId?: string | null;
  /** @nullable */
  memberCount?: number | null;
}

export interface RankInput {
  name: string;
  level: number;
  description?: string;
  color?: string;
  discordRoleId?: string;
}

export interface RankUpdate {
  name?: string;
  level?: number;
  description?: string;
  color?: string;
  discordRoleId?: string;
}

export interface Department {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  discordRoleId?: string | null;
  /** @nullable */
  leaderId?: number | null;
  /** @nullable */
  leaderName?: string | null;
  /** @nullable */
  memberCount?: number | null;
}

export interface DepartmentInput {
  name: string;
  description?: string;
  color?: string;
  discordRoleId?: string;
  leaderId?: number;
}

export interface DepartmentUpdate {
  name?: string;
  description?: string;
  color?: string;
  discordRoleId?: string;
  /** @nullable */
  leaderId?: number | null;
}

export interface PatrolLog {
  id: number;
  memberId: number;
  /** @nullable */
  memberName?: string | null;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  server?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}

export interface PatrolLogInput {
  memberId: number;
  startTime: string;
  endTime?: string;
  durationMinutes?: number;
  server?: string;
  notes?: string;
}

export interface PatrolLogUpdate {
  startTime?: string;
  endTime?: string;
  durationMinutes?: number;
  server?: string;
  notes?: string;
}

export interface Application {
  id: number;
  applicantName: string;
  /** @nullable */
  discordId?: string | null;
  /** @nullable */
  discordUsername?: string | null;
  type: string;
  /** @nullable */
  careerId?: number | null;
  /** @nullable */
  careerTitle?: string | null;
  status: string;
  /** @nullable */
  answers?: string | null;
  /** @nullable */
  reviewedBy?: string | null;
  /** @nullable */
  reviewNotes?: string | null;
  submittedAt: string;
  /** @nullable */
  reviewedAt?: string | null;
}

export interface ApplicationInput {
  applicantName: string;
  discordId?: string;
  discordUsername?: string;
  type: string;
  careerId?: number;
  answers?: string;
}

export interface ApplicationUpdate {
  status?: string;
  reviewedBy?: string;
  reviewNotes?: string;
}

export interface DisciplinaryRecord {
  id: number;
  memberId: number;
  /** @nullable */
  memberName?: string | null;
  type: string;
  reason: string;
  /** @nullable */
  issuedBy?: string | null;
  issuedAt: string;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  notes?: string | null;
  active?: boolean;
}

export interface DisciplinaryInput {
  memberId: number;
  type: string;
  reason: string;
  issuedBy?: string;
  expiresAt?: string;
  notes?: string;
}

export interface DisciplinaryUpdate {
  type?: string;
  reason?: string;
  issuedBy?: string;
  /** @nullable */
  expiresAt?: string | null;
  notes?: string;
  active?: boolean;
}

export interface ActivityRecord {
  memberId: number;
  memberName: string;
  /** @nullable */
  rankName?: string | null;
  /** @nullable */
  departmentName?: string | null;
  totalPatrolHours: number;
  patrolCount?: number;
  /** @nullable */
  lastPatrol?: string | null;
}

export interface WhitelistEntry {
  id: number;
  discordId: string;
  /** @nullable */
  discordUsername?: string | null;
  /** @nullable */
  characterName?: string | null;
  status: string;
  /** @nullable */
  addedBy?: string | null;
  addedAt?: string;
  /** @nullable */
  notes?: string | null;
}

export interface WhitelistInput {
  discordId: string;
  discordUsername?: string;
  characterName?: string;
  status: string;
  addedBy?: string;
  notes?: string;
}

export interface WhitelistUpdate {
  discordUsername?: string;
  characterName?: string;
  status?: string;
  addedBy?: string;
  notes?: string;
}

export interface Career {
  id: number;
  title: string;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  requirements?: string | null;
  isOpen: boolean;
  createdAt?: string;
  /** @nullable */
  applicationCount?: number | null;
}

export interface CareerInput {
  title: string;
  departmentId?: number;
  description?: string;
  requirements?: string;
  isOpen: boolean;
}

export interface CareerUpdate {
  title?: string;
  /** @nullable */
  departmentId?: number | null;
  description?: string;
  requirements?: string;
  isOpen?: boolean;
}

export type DashboardStatsDepartmentBreakdownItem = {
  name: string;
  count: number;
};

export interface DashboardStats {
  totalMembers: number;
  activeMembers: number;
  pendingApplications: number;
  totalPatrolHours: number;
  openPositions: number;
  whitelistedCount: number;
  suspendedMembers?: number;
  departmentBreakdown?: DashboardStatsDepartmentBreakdownItem[];
  recentActivity?: ActivityRecord[];
}

export interface PatrolSummary {
  memberId: number;
  memberName: string;
  totalHours: number;
  patrolCount: number;
  /** @nullable */
  lastPatrol?: string | null;
}

export type ListMembersParams = {
departmentId?: number;
rankId?: number;
status?: string;
search?: string;
};

export type ListPatrolLogsParams = {
memberId?: number;
startDate?: string;
endDate?: string;
};

export type ListApplicationsParams = {
status?: string;
type?: string;
};

export type ListDisciplinaryParams = {
memberId?: number;
type?: string;
};

export type ListActivityParams = {
memberId?: number;
startDate?: string;
endDate?: string;
};

export type ListWhitelistParams = {
status?: string;
search?: string;
};

export type ListCareersParams = {
departmentId?: number;
isOpen?: boolean;
};