# FMS/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Force Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # Patrol Logs
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Submit an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update/review an application
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # Disciplinary Records
  /disciplinary:
    get:
      operationId: listDisciplinary
      tags: [disciplinary]
      summary: List disciplinary records
      parameters:
        - name: memberId
          in: query
          schema: { type: integer }
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of disciplinary records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DisciplinaryRecord"
    post:
      operationId: createDisciplinary
      tags: [disciplinary]
      summary: Create a disciplinary record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplinaryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisciplinaryRecord"
  /disciplinary/{id}:
    patch:
      operationId: updateDisciplinary
      tags: [disciplinary]
      summary: Update a disciplinary record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplinaryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisciplinaryRecord"
    delete:
      operationId: deleteDisciplinary
      tags: [disciplinary]
      summary: Delete a disciplinary record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # Activity
...
  /careers:
    get:
      operationId: listCareers
      tags: [careers]
      summary: List career positions
      parameters:
        - name: departmentId
          in: query
          schema: { type: integer }
        - name: isOpen
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of career positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Career"
    post:
      operationId: createCareer
      tags: [careers]
      summary: Create a career position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Career"
  /careers/{id}:
    patch:
      operationId: updateCareer
      tags: [careers]
      summary: Update a career position
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Career"
    delete:
      operationId: deleteCareer
      tags: [careers]
      summary: Delete a career position
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # Stats
...
components:
  schemas:
...
    Department:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        discordRoleId: { type: ["string", "null"] }
        leaderId: { type: ["integer", "null"] }
        leaderName: { type: ["string", "null"] }
        memberCount: { type: ["integer", "null"] }
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        color: { type: string }
        discordRoleId: { type: string }
        leaderId: { type: integer }
    DepartmentUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        color: { type: string }
        discordRoleId: { type: string }
        leaderId: { type: ["integer", "null"] }

    # Patrol Log schemas
...
    Application:
      type: object
      required: [id, applicantName, type, status, submittedAt]
      properties:
        id: { type: integer }
        applicantName: { type: string }
        discordId: { type: ["string", "null"] }
        discordUsername: { type: ["string", "null"] }
        type: { type: string }
        careerId: { type: ["integer", "null"] }
        careerTitle: { type: ["string", "null"] }
        status: { type: string }
        answers: { type: ["string", "null"] }
        reviewedBy: { type: ["string", "null"] }
        reviewNotes: { type: ["string", "null"] }
        submittedAt: { type: string }
        reviewedAt: { type: ["string", "null"] }
    ApplicationInput:
      type: object
      required: [applicantName, type]
      properties:
        applicantName: { type: string }
        discordId: { type: string }
        discordUsername: { type: string }
        type: { type: string }
        careerId: { type: integer }
        answers: { type: string }
    ApplicationUpdate:
      type: object
      properties:
        status: { type: string }
        reviewedBy: { type: string }
        reviewNotes: { type: string }

    # Disciplinary schemas
    DisciplinaryRecord:
      type: object
      required: [id, memberId, type, reason, issuedAt]
      properties:
        id: { type: integer }
        memberId: { type: integer }
        memberName: { type: ["string", "null"] }
        type: { type: string }
        reason: { type: string }
        issuedBy: { type: ["string", "null"] }
        issuedAt: { type: string }
        expiresAt: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        active: { type: boolean }
    DisciplinaryInput:
      type: object
      required: [memberId, type, reason]
      properties:
        memberId: { type: integer }
        type: { type: string }
        reason: { type: string }
        issuedBy: { type: string }
        expiresAt: { type: string }
        notes: { type: string }
    DisciplinaryUpdate:
      type: object
      properties:
        type: { type: string }
        reason: { type: string }
        issuedBy: { type: string }
        expiresAt: { type: ["string", "null"] }
        notes: { type: string }
        active: { type: boolean }

    # Activity schemas
...
    Career:
      type: object
      required: [id, title, isOpen]
      properties:
        id: { type: integer }
        title: { type: string }
        departmentId: { type: ["integer", "null"] }
        departmentName: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        requirements: { type: ["string", "null"] }
        isOpen: { type: boolean }
        createdAt: { type: string }
        applicationCount: { type: ["integer", "null"] }
    CareerInput:
      type: object
      required: [title, isOpen]
      properties:
        title: { type: string }
        departmentId: { type: integer }
        description: { type: string }
        requirements: { type: string }
        isOpen: { type: boolean }
    CareerUpdate:
      type: object
      properties:
        title: { type: string }
        departmentId: { type: ["integer", "null"] }
        description: { type: string }
        requirements: { type: string }
        isOpen: { type: boolean }

    # Stats schemas
...