# FMS/lib/api-zod/src/generated/types/department.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Force Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Department {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  discordRoleId?: string | null;
  /** @nullable */
  leaderId?: number | null;
  /** @nullable */
  leaderName?: string | null;
  /** @nullable */
  memberCount?: number | null;
}